requirements.txt
PyYAML==5.1
redis==2.10.6
mock==2.0.0
coverage==4.5.1